Move MPS tests to same file
