+++ engines/pytorch/pytorch-engine/src/test/java/ai/djl/pytorch/integration/MpsTest.java
@@ -13,6 +13,7 @@
 package ai.djl.pytorch.integration;
 
 import ai.djl.Device;
+import ai.djl.modality.Classifications;
 import ai.djl.ndarray.NDArray;
 import ai.djl.ndarray.NDManager;
 import ai.djl.ndarray.types.Shape;
@@ -20,6 +21,9 @@ import ai.djl.ndarray.types.Shape;
 import org.testng.Assert;
 import org.testng.SkipException;
 import org.testng.annotations.Test;
+
+import java.util.Arrays;
+import java.util.List;
 
 public class MpsTest {
 
@@ -36,4 +40,42 @@ public class MpsTest {
             Assert.assertEquals(array.getDevice().getDeviceType(), "mps");
         }
     }
+
+    private static boolean checkMpsCompatible() {
+        return "aarch64".equals(System.getProperty("os.arch")) &&
+                System.getProperty("os.name").startsWith("Mac");
+    }
+
+    @Test
+    public void testToTensorMPS() {
+        if (!checkMpsCompatible()) {
+            throw new SkipException("MPS toTensor test requires Apple Silicon macOS.");
+        }
+
+        // Test that toTensor does not fail on MPS (e.g. due to use of float64 for division)
+        try (NDManager manager = NDManager.newBaseManager(Device.fromName("mps"))) {
+            NDArray array = manager.create(127f).reshape(1, 1, 1, 1);;
+            NDArray tensor = array.getNDArrayInternal().toTensor();
+            Assert.assertEquals(tensor.toFloatArray(), new float[]{127f/255f});
+        }
+    }
+
+    @Test
+    public void testClassificationsMPS() {
+        if (!checkMpsCompatible()) {
+            throw new SkipException("MPS classification test requires Apple Silicon macOS.");
+        }
+
+        // Test that classifications do not fail on MPS (e.g. due to conversion of probabilities to float64)
+       try (NDManager manager = NDManager.newBaseManager(Device.fromName("mps"))) {
+            List<String> names = Arrays.asList("First", "Second", "Third", "Fourth", "Fifth");
+            NDArray tensor = manager.create(new float[]{0f, 0.125f, 1f, 0.5f, 0.25f});
+            Classifications classifications = new Classifications(
+                    names,
+                    tensor
+            );
+            Assert.assertNotNull(classifications.topK(1).equals(Arrays.asList("Third")));
+        }
+    }
+
 }
